+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Check TGW state before attempting to create attachment
+
 ## [0.1.5] - 2022-09-16
 
 ### Fixed